feat: refactor export button component and update import path
